Refactor parser for better readability and structure
- Extracted recursion depth checking to `src/parser/recursion.rs`
- Refactored `build_statement` to use helper function `build_regular_statement`
- Simplified `parse_method_parameters` loop to avoid nested `if`s and `clippy` suppression
- Updated `.jules/forge.md` with new learnings
